src/requirements.txt
bitcoin==1.1.42
certifi==2017.7.27.1
chardet==3.0.4
coinbase==2.0.6
cytoolz==0.8.2
ethereum==2.0.4
ethereum-abi-utils==0.4.0
ethereum-client-utils==0.3.2
ethereum-rpc-client==0.4.4
ethereum-utils==0.4.0
future==0.16.0
idna==2.6
pbkdf2==1.3
peewee==2.10.1
psycopg2==2.7.3.1
pycrypto==2.6.1
pycryptodome==3.4.7
pyethash==0.1.27
pylru==1.0.9
pysha3==1.0.2
python-telegram-bot==8.0
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rlp==0.6.0
scrypt==0.8.0
six==1.10.0
toolz==0.8.2
urllib3==1.22
web3==3.13.4
